Complexity calculation moved to enum
